Refactoring Domain Package : Category, News, tag
