+++ src/main/java/com/snack/news/domain/base/BaseTimeEntity.java
@@ -1,20 +1,18 @@
-package com.snack.news.domain;
+package com.snack.news.domain.base;
 
 import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
+import javax.persistence.EntityListeners;
+import javax.persistence.MappedSuperclass;
+import java.time.LocalDateTime;
+
 @Getter
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener.class)
-abstract class BaseTimeEntity {
+public abstract class BaseTimeEntity {
 	@CreatedDate
 	private LocalDateTime createAt;
 	@LastModifiedDate
