+++ src/main/java/com/snack/news/domain/tag/Tag.java
@@ -1,4 +1,4 @@
-package com.snack.news.domain;
+package com.snack.news.domain.tag;
 
 
 import lombok.Builder;
