+++ src/main/java/com/snack/news/repository/CategoryRepository.java
@@ -1,8 +1,7 @@
 package com.snack.news.repository;
 
+import com.snack.news.domain.category.Category;
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.snack.news.domain.Category;
 
 public interface CategoryRepository extends JpaRepository<Category, Long> {
 }
