+++ src/test/java/com/snack/news/fixture/NewsFixture.java
@@ -1,6 +1,6 @@
 package com.snack.news.fixture;
 
-import com.snack.news.domain.News;
+import com.snack.news.domain.news.News;
 import com.snack.news.dto.NewsDto;
 import org.junit.Before;
